📝 (ChatsController.java): Add a dynamic filter to only accept chats that are in the whiteListedChats list
TdlibChatListSlice.java: Remove hardcoded whiteListedChats and implement a dynamic filter based on the list of whiteListedChats
The hardcoded whiteListedChats list has been removed from TdlibChatListSlice.java and replaced with a dynamic filter that checks if the chat id is in the whiteListedChats list. In ChatsController.java, a dynamic filter has been added to only accept chats that are in the whiteListedChats list, improving maintainability and flexibility in managing the list of allowed chats.